SWATCH-4805: renaming classes to implicitly communicate hierarchy
